initial progress on fixing tests

+++ Core/Services/StitchCoreRemoteMongoDBService/Tests/StitchCoreRemoteMongoDBServiceTests/Sync/XCMongoMobileTestCase.swift
@@ -73,16 +73,20 @@ class TestEventDelegate: ChangeEventDelegate {
     }
 }
 
-private class DeinitializingDataSynchronizer: DataSynchronizer {
+private class TestCaseDataSynchronizer: DataSynchronizer {
     private let deinitializingInstanceKey: String
-    private let localClient: MongoClient
+//    let localClient: MongoClient
+    private let deinitializing: Bool
 
-    init(instanceKey: String,
+    init(deinitializing: Bool,
+         instanceKey: String,
          coreRemoteMongoClient: CoreRemoteMongoClient,
          localClient: MongoClient) throws {
+        self.deinitializing = deinitializing
         self.deinitializingInstanceKey = instanceKey
-        self.localClient = localClient
+//        self.localClient = pr
 
+        print("initializing a test case data synchronizer")
         let mockServiceClient = MockCoreStitchServiceClient.init()
         try super.init(
             instanceKey: instanceKey,
@@ -94,10 +98,12 @@ private class DeinitializingDataSynchronizer: DataSynchronizer {
     }
 
     deinit {
-        try? self.localClient.db(
-            DataSynchronizer
-                .localConfigDBName(withInstanceKey: deinitializingInstanceKey)
-            ).drop()
+        if deinitializing {
+            try? self.localClient.db(
+                DataSynchronizer
+                    .localConfigDBName(withInstanceKey: deinitializingInstanceKey)
+                ).drop()
+        }
     }
 }
 
@@ -120,10 +126,57 @@ class XCMongoMobileTestCase: XCTestCase {
 
     private(set) var mockServiceClient: MockCoreStitchServiceClient!
     private(set) var coreRemoteMongoClient: CoreRemoteMongoClient!
-    lazy var dataSynchronizer: DataSynchronizer =
-        try! DeinitializingDataSynchronizer.init(instanceKey: instanceKey.oid,
-                                                 coreRemoteMongoClient: self.coreRemoteMongoClient,
-                                                 localClient: localClient)
+
+    private var storedDataSynchronizer: DataSynchronizer!
+    var dataSynchronizer: DataSynchronizer! {
+        get {
+            if storedDataSynchronizer == nil {
+                storedDataSynchronizer = try! TestCaseDataSynchronizer.init(
+                    deinitializing: true,
+                    instanceKey: instanceKey.oid,
+                    coreRemoteMongoClient: self.coreRemoteMongoClient,
+                    localClient: localClient)
+            }
+            return storedDataSynchronizer
+        }
+        set {
+            storedDataSynchronizer = newValue
+        }
+    }
+
+    func replaceDataSynchronizer(
+        deinitializing: Bool,
+        withUndoDocuments undoDocuments: [Document] = []
+    ) throws {
+        dataSynchronizer.stop()
+        dataSynchronizer = nil
+
+        if undoDocuments.isEmpty {
+            print("replacing data synchronizer with NO undo documents")
+        } else {
+            print("replacing data synchronizer with SOME undo documents")
+        }
+
+        try undoDocuments.forEach { doc in
+            try undoCollection().insertOne(doc)
+        }
+
+        dataSynchronizer = try! TestCaseDataSynchronizer.init(
+            deinitializing: deinitializing,
+            instanceKey: instanceKey.oid,
+            coreRemoteMongoClient: self.coreRemoteMongoClient,
+            localClient: localClient
+        )
+
+        // perform a no-op write so that we wait for the recovery pass to complete. This works
+        // since the recovery routine write-locks all namespaces until recovery is done.
+        _ = try dataSynchronizer.updateOne(
+            filter: ["_id": "nonexistent"],
+            update: ["$set": ["a": 1] as Document],
+            options: nil,
+            in: namespace
+        )
+    }
 
     private var _instanceKey = ObjectId()
     open var instanceKey: ObjectId {
@@ -171,8 +224,19 @@ class XCMongoMobileTestCase: XCTestCase {
                         withCollectionType: type)
     }
 
+    func isUndoCollectionEmpty() throws -> Bool {
+        return try undoCollection().count() == 0
+    }
+
+    func undoCollection() throws -> MongoCollection<Document> {
+        return try dataSynchronizer.undoCollection(for: namespace)
+    }
+
     func localCollection() throws -> MongoCollection<Document> {
-        return try localCollection(for: namespace)
+        return try localCollection(for: MongoNamespace.init(
+            databaseName: DataSynchronizer.localUserDBName(withInstanceKey: instanceKey.oid, for: namespace),
+            collectionName: namespace.collectionName
+        ))
     }
 
     func localCollection(for namespace: MongoNamespace) throws -> MongoCollection<Document> {
